Make delete() non blocking
The core machine controller will not call us back on error or an
explicit request to re-queue by returning RequeueAfterError.

By the time delete is called the machine object has gone. If were to
call machine.Update() and add an annotation for the operation.Id (so
that we could keep track of it in follow-up calls) the Update() will
fail.

Ideally we would like to return nil when the cloud operation is
"DONE". However, the current design of the core controller does not
allow for an idempotent delete operation.