Disable keep alives on termination handler connection
The GCP metadata service keeps the connnection alive, this means that if 
we try to intercept calls to the metadata service after the connection 
is established, we have to restart the process.

We use this interception technique during the E2E suite to verify that 
the termination handler will respond approriately to a termination event 
if observed.